Inline load_trajectory_final_output into orchestrator.py
Only used by the orchestrator, so a separate trajectory.py module was
unnecessary. Deletes trajectory.py.
